Remove multi-head support from strq, move into purger.
The multi-head idea was efficient for multiple parallel dequeuers,
but at this point we're more concerned about layer races.
Switching the strq code back to single-head is a first step
towards serializing the multiple purgers (varnishes) so that the
first must complete a purge before it's handed to the next (and
future queue delays as well).

This leaves the codebase temporarily-disfunctional in that only
the first purger now gets any requests.  To be fixed in the
followups as the refactor continues.

Also, there's no need to pass and utilize strq+purger in parallel
from everywhere.  Logically at this point, one purger "owns" its
own strq, so hide it within as an implementation detail of the
purger.

Change-Id: Id30875ecec7fdef09f12402f2dba2fea64410538