Chain the purgers together and split their stats
This completes the ideas that began in the earlier "Remove
multi-head support from strq" commit.  It hooks up the
commandline-specified purgers in a chain, such that the URLs
received via HTCP flow like so:

HTCP -> Receiver (parse, generate HTTP PURGE req)
  -> Purger0_Queue -> Purge0_Varnish (confirm response)
    -> Purger1_Queue -> Purge1_Varnish (confirm response)
      -> Purger2_Queue ...

This ensures purging happens serially according to the order the
purge addresses are specified on the commandline, with one purger
claiming completion over HTTP before the next even gets handed the
request to enqueue for itself, at the cost of multiple queues and
slight delays.

The statistics output is now split up accordingly as well, but is
due for further refactoring later that will change it
more-substantially...

Change-Id: I0f03a07ddc89f11f6322d4f20ece6b9b7c87b785