Refactor (rewrite?!) purging code
This refactor removes the PST_NOTCONN_WAIT state
(disconnected+idle) from the state machine.  The code no longer
has such a state as it always reconnects immediately, even if
there is no currently-pending output.  This also kills the
idle_timeout parameter (which controlled falling from
connected+idle to disconnected+idle).

Much of the rest of the logic was de-duplicated into a few new
helper functions, and some of the internal state magic was made a
bit more idempotent and efficient.

Also, "purger_t* s" was changed to "purger_t* p" throughout to
reduce confusion (was a leftover from when "purger" was called
"sender")

Change-Id: Icd5e96a1bae4dac65c81f80b07f99e9388608fe0