Set q_mem + q_max_mem stats correctly at startup
Very minor bugfix, not worth a separate release.

The q_mem and q_max_mem stats are left at the value "0" on
startup, and do not get bumped to correct near-initial values
until the first purge comes into each queue.  Not really
noticeable in pragmatic cases, but it looks "odd" in stats on
cache_misc, which receives no purge traffic currently and thus
always shows zeros when the real value should be 24576 bytes in
both fields.

Change-Id: I5475a7b3bc65c7b80d7dab62bb0d21c170275ff1